Out-of-order processing of CLOSE_CONN_S messages
When the client closes the connection, it is useless to transfer any more data.
Such data could only be dropped on the local side of the proxy as the connection
to forward it to no longer exists.

Out-of-order processing of CLOSE_CONN_C message must not be done. That would
cut transfers short.